+++ BTDB/IOC/IContainer.cs
@@ -7,5 +7,8 @@ namespace BTDB.IOC
         object Resolve(Type type);
         object ResolveNamed(string name, Type type);
         object ResolveKeyed(object key, Type type);
+        object? ResolveOptional(Type type);
+        object? ResolveOptionalNamed(string name, Type type);
+        object? ResolveOptionalKeyed(object key, Type type);
     }
 }